// src/app/models/character.model.ts
export interface Character {
    name: string;
    image: string;
    id: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color:string;

    eye_color:string;
    birth_year:string;
    gender:string;
    homeworld: string;
    films:string[];
    species:string[],
    vehicles: string[]
    starships: string[];
    url:string
}